+++ Twinder/DeepLinks.swift
@@ -10,6 +10,13 @@ import Foundation
 import UIKit
 
 func twtrScreenName(screen_name: String){
+    let allowedScreenNameCharacters = NSCharacterSet(charactersInString: "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789_")
+    let screenNameLength = (screen_name as NSString).length
+    if screenNameLength == 0 || screenNameLength > 15 ||
+        screen_name.rangeOfCharacterFromSet(allowedScreenNameCharacters.invertedSet) != nil {
+        return
+    }
+
     let components = NSURLComponents()
     components.scheme = "twitter"
     components.host = "user"